[XTENSA] Include complete AR register file in xtensa_gregset_t
Increate the xtensa_gregset_t structure (elf gregset) from 32 to 128.
This allows us to include the complete AR register file and has plenty
of space for future extentions. This patch also adds the windowstart
register back into this structure to define the valid frames in the
register file. Windowbase is assumed to be 0.

Signed-off-by: Chris Zankel <chris@zankel.net>