[XTENSA] Fix coprocessor context switching
The coprocessor states of all coprocessors were saved to
the same location. We were also using the wrong value for
in determining the coprocessors used by the current thread.
The cpenable field in the thread_info structure that was used
before, is only updated on a task-switch, so for the current
task, we have to use the cpenable register.

Signed-off-by: Dan Nicolaescu <dann@xtensa-linux.org>
Signed-off-by: Marc Gauthier <marc@tensilica.com>
Signed-off-by: Chris Zankel <chris@zankel.net>