[XTENSA] Spinlock corrections plus documentation
Correct the __raw_write_can_lock macro.  It's logic was backward.

In __raw_read_lock, re-order instructions for minor optimization and
improved clarity.  That is, the addi is really not relevant to the
bltz.  The addi is irrelevant only because of the way we chose to
define valid lock values (something not obvious at first glance).

In __raw_read_trylock, re-order the addi for the same reason as above.
Also, change the return condition to achieve multiple readers (i.e.,
the previous condition did not provide for multiple readers, which is
the primary advantage of the rwlock).

Document the spinlock and rwlock approach on Xtensa.

Document some return values.

Signed-off-by: Joe Taylor <joe@tensilica.com>