[XTENSA] - Major update to support memory mapped files correctly.
mmap() problem with pthreads and the LTP open posix test suite
indicated kernel problems. Files were appearing non-linear due
to the pte_file() inline function not working correctly.

Marc and I thought the cause was the table not being
correct and clear so we cleaned it up and updated it to work correctly.

Another problem was the default vma protection bits were wrong. The
LTP test was making the page unreadable and that didn't allow the pte
to appear to be present; so was added a new PRESENT bit next to the
software bits.

This fix is still a bit experimental, some results still look suspicious,
but mmap() files seem fine now.

Signed-off-by: Pete Delaney <piet@tensilica.com>