[XTENSA] Fixed memmove(), bcopy(), and memcpy().
1. Atheros modules use memmove() with overlapping
   src and dst. This fix to memmove makes it handle
   overlaping src and src using a C library from the
   uClibc library. Performance loss over asm code is
   5% for aligned data and 75% for unaligned data
   for test moving data both forward and backward
   with random overlapping from 1 to 64 bytes.

2. Fixed memcpy() to use unsigned arithmetic for
   addresses comparisons.

3. Modified bcopy() to call memmove(). It's suppose to
   handle overlapping src and data like memmove(). It's
   hardly used, just IPV6 interface macro possibly.

memmove.c is using a local memcopy.h file from uClibc.
It likely belongs in ../include/asm/ but I'll try come
up with an assembler version of memmove() in the not to
distant future and we will then drop memmove.c and it's
associated memcopy.h from uClibc.

LTP test seem to be running fine with the new code, thought
I haven't go results for long test yet.

Signed-off-by: Pete Delaney <piet@tensilica.com>