xtensa: put stack pointer into a1 for call12 spill case
call12 spill procedure starts with the following code:

	s32e    a4, a1, -16     # a1 is valid with an empty spill area
	l32e    a4, a5, -12
	s32e    a8, a4, -48
	mov     a8, a4
	l32e    a4, a1, -16
	j       .Lc12c

a1 is expected to point to the topmost stack frame, but this is almost
never the case after the initial register window rotation. The result is
either stack corruption or infinite attempt to call fixup routine after the
first s32e.

This may be reproduced by invoking spill syscall after the following
sequence of calls: 4 -> 12 -> 12 -> 8 on configuration with 32 physical
registers.

The rest of the patch fixes the caller of _spill_registers, saving/restoring
additional registers across the call.

Cc: stable@vger.kernel.org
Signed-off-by: Max Filippov <jcmvbkbc@gmail.com>