WIP: xtensa: add VECTORS_DATA
XTENSA_VECTORS_SEPARATE switch allows placing vectors in a dedicated
region, presumably IRAM, where memory access is fast. This is not
sufficient for the fast exception handling as most of the vectors need
exception handlers table and some stack. Collect data needed for fast
exception handlers in a new section and place it at specific location
when vector separation is enabled.

Signed-off-by: Max Filippov <jcmvbkbc@gmail.com>