WIP: mm/nommu: don't require sharing on XIP kernel for ptraced process
This fixes an attempt to load executable code segment instead of
mmapping it directly in a process running under strace. On xtensa kernel
built with XIP for a platform where RAM is not executable this results
in an immediate fault.

TODO: this solution is working but a thought about a bigger picture is
needed.

Signed-off-by: Max Filippov <jcmvbkbc@gmail.com>