mtd/map_esp32: fix race at initialization
There's a race between receiving a response from firmware to a request
for comm structure allocation and a partition table parsing. If
partition table parsing request comes first it gets an error which
results in empty partition table and failure to boot.
Wait in the probe function until comm structure allocation is complete.

Signed-off-by: Max Filippov <jcmvbkbc@gmail.com>